+++ R/atp_plot.R
@@ -3,6 +3,8 @@
 #' Generate the ATP Plot
 #' @param energetics A table of calculated glycolysis and OXPHOS rates.
 #' Returned by `get_energetics`
+#' @param model The linear model used to estimate mean and confidence
+#' intervals: ordinary least squares (`"ols"`) or mixed-effects (`"mixed"`)
 #' @param error_bar Whether to plot error bars as standard deviation (`"sd"`)
 #' or confidence intervals (`"ci"`)
 #' @param conf_int The confidence interval percentage. Should be between 0 and 1
@@ -11,9 +13,16 @@
 #' @param basal_vs_max Whether to plot `"basal"` or `"max"` respiration
 #' @param glyc_vs_resp Whether to plot glycolysis (`"glyc"`)  or respiration (`"resp"`)
 #' @param group_label Label for the experimental group to populate the legend title
+#' @param sep_reps Whether to calculate summary statistics on the groups with
+#' replicates combined. The current default `FALSE` combines replicates, but
+#' future releases will default to `TRUE` providing replicate-specific
+#' summaries.
+#' @param ci_method The method used to compute confidence intervals for the
+#' mixed-effects model: `"Wald"`, `"profile"`, or `"boot"` passed to
+#' `lme4::confint.merMod()`.
 #' @return a ggplot
 #'
-#' @importFrom ggplot2 ggplot aes geom_point labs xlab ylab geom_linerange xlim ylim geom_crossbar theme_bw .data
+#' @importFrom ggplot2 ggplot aes geom_point labs xlab ylab geom_linerange xlim ylim geom_linerange theme_bw .data position_dodge2
 #' @export
 #'
 #' @details
@@ -27,31 +36,55 @@
 #'   list.files(pattern = "*.xlsx", full.names = TRUE)
 #' seahorse_rates <- read_data(rep_list, sheet = 2)
 #' partitioned_data <- partition_data(seahorse_rates)
-#' energetics <- get_energetics(partitioned_data, ph = 7.4, pka = 6.093, buffer = 0.1)
-#' atp_plot(energetics)
+#' energetics <- get_energetics(
+#'   partitioned_data,
+#'   ph = 7.4,
+#'   pka = 6.093,
+#'   buffer = 0.1
+#' )
+#' atp_plot(energetics, sep_reps = FALSE)
 #'
-#' atp_plot(energetics, basal_vs_max = "max")
+#' atp_plot(energetics, basal_vs_max = "max", sep_reps = FALSE)
 #'
-#' atp_plot(energetics, basal_vs_max = "basal", glyc_vs_resp = "resp")
-#'
+#' atp_plot(
+#'   energetics,
+#'   basal_vs_max = "basal",
+#'   glyc_vs_resp = "resp",
+#'   sep_reps = TRUE
+#' )
 #' # to change fill, the geom_point shape number should be between 15 and 25
-#' atp_plot(energetics, shape = 21) + # filled circle
-#'   ggplot2::scale_fill_manual(values = c("#e36500", "#b52356", "#3cb62d", "#328fe1"))
+#' atp_plot(
+#'   energetics,
+#'   sep_reps = FALSE
+#' ) +
+#'   ggplot2::scale_fill_manual(
+#'     values = c("#e36500", "#b52356", "#3cb62d", "#328fe1")
+#'   )
 #'
 #' # to change color, use ggplot2::scale_color_manual
-#' atp_plot(energetics) +
-#'   ggplot2::scale_color_manual(values = c("#e36500", "#b52356", "#3cb62d", "#328fe1"))
+#' atp_plot(energetics, sep_reps = FALSE) +
+#'   ggplot2::scale_color_manual(
+#'     values = c("#e36500", "#b52356", "#3cb62d", "#328fe1")
+#'   )
 atp_plot <- function(
     energetics,
+    model = "ols",
     error_bar = "ci",
     conf_int = 0.95,
     size = 2,
-    shape = 21,
+    shape = 16,
     basal_vs_max = "basal",
     glyc_vs_resp = "glyc",
-    group_label = "Experimental group") {
+    group_label = "Experimental group",
+    sep_reps = FALSE,
+    ci_method = "Wald") {
   # Sanity checks
   stopifnot("'error_bar' should be 'sd' or 'ci'" = error_bar %in% c("sd", "ci"))
+  stopifnot("'model' should be 'ols' or 'mixed'" = model %in% c("ols", "mixed"))
+  stopifnot(
+    "cannot run mixed-effects model with `sep_reps = TRUE`" =
+      (model == "mixed" & !sep_reps) | (model == "ols")
+  )
   stopifnot("'conf_int' should be between 0 and 1" = conf_int > 0 && conf_int < 1)
 
   data_cols <- c(
@@ -60,7 +93,7 @@ atp_plot <- function(
     "ATP_basal_resp",
     "ATP_max_resp"
   )
-
+  replicate <- NULL
   missing_cols <- setdiff(data_cols, colnames(energetics))
   if (length(missing_cols) != 0) {
     stop(paste0("'", missing_cols, "'", " column was not found in input data\n"))
@@ -69,10 +102,16 @@ atp_plot <- function(
   # suppress "no visible binding for global variable" error
   exp_group <- NULL
 
+  # TODO: make sep_reps = TRUE the default
+  multi_rep <- length(unique(energetics$replicate)) > 1
+  if (!sep_reps && missing(sep_reps) && multi_rep) warning(sep_reps_warning)
+
   energetics_summary <- get_energetics_summary(
     energetics,
+    model = model,
     error_metric = error_bar,
-    conf_int = conf_int
+    conf_int = conf_int,
+    sep_reps = sep_reps
   )
   # Identify numeric columns and replace their negative values with 0
   numeric_cols <- names(energetics_summary)[sapply(energetics_summary, is.numeric)]
@@ -104,24 +143,26 @@ atp_plot <- function(
     energetics_summary,
     aes(
       x = exp_group, y = .data[[data_column.mean]],
-      exp_group,
-      color = exp_group,
-      fill = exp_group
+      color = if (sep_reps && multi_rep) replicate else NULL
     )
   ) +
-    geom_point(size = size, shape = shape) +
-    geom_crossbar(
+    geom_point(
+      size = size,
+      shape = shape,
+      position = position_dodge2(width = 0.3)
+    ) +
+    geom_linerange(
       aes(
         x = exp_group, y = .data[[data_column.mean]],
         ymin = .data[[data_column.lower_bound]],
         ymax = .data[[data_column.higher_bound]],
       ),
-      alpha = 0.5,
-      data = energetics_summary
+      data = energetics_summary,
+      position = position_dodge2(width = 0.3)
     ) +
     xlab(group_label) +
     ylab("ATP Production (JATP)") +
     labs(title = paste0("ATP Production from ", basal_vs_max_label, " ", glyc_vs_resp_label)) +
-    labs(color = "Experimental Group", fill = "Experimental Group") +
+    labs(color = "Replicate") +
     theme_bw()
 }